Rename vtkInterface module to
